\name{lifetech}
\docType{data}
\alias{lifetech}
\title{The processed data generated using the LifeTech software.}
\description{These data are also used to demonstrate the functionality
  of the miRcomp package.}
\usage{data(lifetech)}
\format{A list with two elements.
  \tabular{ll}{
  ct \tab a miRNA x sample matrix of expression estimates \cr
  qc \tab a miRNA x sample matrix of quality scores (here AmpScores)
  }}
\keyword{datasets}
\examples{
  data(lifetech)
}
