\name{miRcompShinyApp}
\alias{miRcompShinyApp}
\title{Launch a Shiny app to use the miRcomp package interactively}
\description{
  This function launches the miRcomp Shiny app locally.
}
\usage{
miRcompShinyApp()
}
\author{Matthew N. McCall and Lauren Kemperman}
\examples{
  ## miRcompShinyApp()
}
\keyword{manip}

