\name{qpcRb4}
\docType{data}
\alias{qpcRb4}
\title{The processed data generated using the 4 parameter sigmoidal method from the
  qpcR software.}
\description{These data are also used to demonstrate the functionality
  of the miRcomp package.}
\usage{data(qpcRb4)}
\format{A list with two elements.
  \tabular{ll}{
  ct \tab a miRNA x sample matrix of expression estimates \cr
  qc \tab a miRNA x sample matrix of quality scores (here R-squared values)
  }}
\keyword{datasets}
\examples{
  data(qpcRb4)
}
