\name{qpcRb5}
\docType{data}
\alias{qpcRb5}
\title{The processed data generated using the 5 parameter sigmodial method from the
  qpcR software.}
\description{These data are also used to demonstrate the functionality
  of the miRcomp package.}
\usage{data(qpcRb5)}
\format{A list with two elements.
  \tabular{ll}{
  ct \tab a miRNA x sample matrix of expression estimates \cr
  qc \tab a miRNA x sample matrix of quality scores (here R-squared values)
  }}
\keyword{datasets}
\examples{
  data(qpcRb5)
}
