\name{qpcRcm3}
\docType{data}
\alias{qpcRcm3}
\title{The processed data generated using the mechanistic cm3 method
  proposed by Carr and Moore (2012) implemented in the
  qpcR software package.}
\description{These data are also used to demonstrate the functionality
  of the miRcomp package.}
\usage{data(qpcRcm3)}
\format{A list with two elements.
  \tabular{ll}{
  ct \tab a miRNA x sample matrix of expression estimates \cr
  qc \tab a miRNA x sample matrix of quality scores (here R-squared values)
  }}
\keyword{datasets}
\examples{
  data(qpcRcm3)
}
