\name{qpcRdefault}
\docType{data}
\alias{qpcRdefault}
\title{The processed data generated using the default version of the
  qpcR software.}
\description{These data are also used to demonstrate the functionality
  of the miRcomp package.}
\usage{data(qpcRdefault)}
\format{A list with two elements.
  \tabular{ll}{
  ct \tab a miRNA x sample matrix of expression estimates \cr
  qc \tab a miRNA x sample matrix of quality scores (here R-squared values)
  }}
\keyword{datasets}
\examples{
  data(qpcRdefault)
}
