\name{qpcRl5}
\docType{data}
\alias{qpcRl5}
\title{The processed data generated using the 5 parameter log-logistic method from the
  qpcR software.}
\description{These data are also used to demonstrate the functionality
  of the miRcomp package.}
\usage{data(qpcRl5)}
\format{A list with two elements.
  \tabular{ll}{
  ct \tab a miRNA x sample matrix of expression estimates \cr
  qc \tab a miRNA x sample matrix of quality scores (here R-squared values)
  }}
\keyword{datasets}
\examples{
  data(qpcRl5)
}
