\name{qpcRlinexp}
\docType{data}
\alias{qpcRlinexp}
\title{The processed data generated using the linear-exponential method
  implemented in the qpcR software package.}
\description{These data are also used to demonstrate the functionality
  of the miRcomp package.}
\usage{data(qpcRlinexp)}
\format{A list with two elements.
  \tabular{ll}{
  ct \tab a miRNA x sample matrix of expression estimates \cr
  qc \tab a miRNA x sample matrix of quality scores (here R-squared values)
  }}
\keyword{datasets}
\examples{
  data(qpcRlinexp)
}
